refactor: extract AppState+ScreenshotCapture (4 delegators) (#615, #601 slice G)
Seventh slice of #601. Creates new AppState+ScreenshotCapture.swift with 4
screenshotCaptureController delegators (1 method + 3 computed properties):

- captureScreenshot() — async method
- prepareForScreenshotSelection — get/set property pair
- restoreAfterScreenshotSelection — get/set property pair
- isScreenshotCaptureInProgress — read-only computed property

Non-adjacent 3-block extraction: property pairs at 60-67, isCaptureInProgress
at 633-635, captureScreenshot at 916-918. Reverse-order sed per #556.

Slice G expanded from ticket's 2-property scope to 4 members after discovering
2 more same-collaborator delegators via grep (captureScreenshot method +
isScreenshotCaptureInProgress property) — coherent 4-member domain.

Byte-identical moves; SHAs verified against origin/main:
- prop pairs (60-67): cff85bd3ba274037748ee256342c1b7c5e3500ae7c458cf7b8145a60696eb31f
- isScreenshotCaptureInProgress (632-634): c1e320fb806441916793e6e8881efceacadbb8d6176b45ab1fdc5ac961449e4b
- captureScreenshot (916-918): f56c7e3a066cebe0b4565b4ee732ca45083cb6a77aab6a25a8f70bd0463ebe94

Safety checklist:
- All delegator tests pass (single-expression bodies, no side effects, no branching)
- $binding sanity checks all zero
- Observation forwarding preserved
- Green tests: 667
